repo_results/run-20240308-125808_LLMatic-ZeroShot-CodeT-SelfRefine-Reflexion-ZeroShot_donnemartin-system-design-oop-url_shortener-donnemartin-system-design-oop-twitter-donnemartin-system-design-oop-whatsapp_0_20-runs_log/twitter/7/LLMatic/requirements.txt
Flask==1.1.2
pytest==6.2.2
Werkzeug==1.0.1
PyJWT==1.7.1
